Split secure and unsecure sandboxes into two classes in preparation of further divergence in implementation

+++ worker/sandbox.py
@@ -19,8 +19,28 @@ except ImportError:
 class SandboxError(StandardError):
     pass
 
-class _Jail(object):
-    def __init__(self):
+def _monitor_file(fd, q):
+    while True:
+        line = fd.readline()
+        if not line:
+            q.put(None)
+            break
+        q.put(line.rstrip('\r\n'))
+
+class Jail(object):
+    """ Provide a secure sandbox to run arbitrary commands in.
+
+    This will only function on specially prepared Ubuntu systems.
+
+    """
+    def __init__(self, working_directory):
+        """Initialize a new sandbox for the given working directory.
+
+        working_directory: the directory in which the shell command should
+                           be launched. Files from this directory are copied
+                           into the secure space before the shell command is
+                           executed.
+        """
         self.locked = False
         jail_base = "/srv/chroot"
         all_jails = os.listdir(jail_base)
@@ -46,12 +66,36 @@ class _Jail(object):
         self.chroot_cmd = "sudo -u {0} schroot -u {0} -c {0} -d {1} -- ".format(
                 self.name, "/home/jailuser")
 
+        self._is_alive = False
+        self.command_process = None
+        self.stdout_queue = Queue()
+        self.stderr_queue = Queue()
+        self._prepare_with(working_directory)
+
     def __del__(self):
         if self.locked:
             raise SandboxError("Jail object for %s freed without being released"
                     % (self.name))
 
+    @property
+    def is_alive(self):
+        """Indicates whether a command is currently running in the sandbox"""
+        if self._is_alive:
+            sub_result = self.command_process.poll()
+            if sub_result is None:
+                return True
+            self._is_alive = False
+        return False
+
     def release(self):
+        """Release the sandbox for further use
+
+        Unlocks and releases the jail for reuse by others.
+        Must be called exactly once after Jail.is_alive == False.
+
+        """
+        if self.is_alive:
+            raise SandboxError("Sandbox released while still alive")
         if not self.locked:
             raise SandboxError("Attempt to release jail that is already unlocked")
         lock_dir = os.path.join(self.base_dir, "locked")
@@ -67,7 +111,7 @@ class _Jail(object):
         os.rmdir(lock_dir)
         self.locked = False
 
-    def prepare_with(self, command_dir):
+    def _prepare_with(self, command_dir):
         if os.system("%s c %d" % (self.jchown, self.number)) != 0:
             raise SandboxError("Error returned from jail_own c %d in prepare"
                     % (self.number,))
@@ -87,101 +131,26 @@ class _Jail(object):
         self.command_dir = command_dir
 
     def retrieve(self):
+        """Copy the working directory back out of the sandbox."""
+        if self.is_alive:
+            raise SandboxError("Tried to retrieve sandbox while still alive")
         os.system("rm -rf %s" % (self.command_dir,))
         if os.system("%s c %d" % (self.jchown, self.number)) != 0:
             raise SandboxError("Error returned from jail_own c %d in prepare"
                     % (self.number,))
         os.system("cp -r %s %s" % (self.home_dir, self.command_dir))
 
-    def signal(self, signal):
-        if not self.locked:
-            raise SandboxError("Attempt to send %s to unlocked jail" % (signal,))
-        result = subprocess.call("sudo -u {0} kill -{1} -1".format(
-            self.name, signal), shell=True)
-        if result != 0:
-            raise SandboxError("Error returned from jail %s sending signal %s"
-                    % (self.name, signal))
-
-    def kill(self):
-        self.signal("KILL")
-        self.signal("CONT")
-
-    def pause(self):
-        self.signal("STOP")
-
-    def resume(self):
-        self.signal("CONT")
-
-def _monitor_file(fd, q):
-    while True:
-        line = fd.readline()
-        if not line:
-            q.put(None)
-            break
-        q.put(line.rstrip('\r\n'))
-
-class Sandbox:
-    """Provide a sandbox to run arbitrary commands in.
-
-    The sandbox class is used to invoke arbitrary shell commands. Its main
-    feature is that it has the option to launch the shell command inside a
-    jail, in order to totally isolate the command.
-
-    """
-
-    def __init__(self, working_directory, secure=None):
-        """Initialize a new sandbox and invoke the given shell command inside.
-
-        working_directory: the directory in which the shell command should
-                           be launched. If security is enabled, files from
-                           this directory are copied into the VM before the
-                           shell command is executed.
-        secure: really use a jail or just run the command directly
-                defaults to True when a server_info module is found, False
-                otherwise
-
-        """
-        self._is_alive = False
-        self.command_process = None
-        self.stdout_queue = Queue()
-        self.stderr_queue = Queue()
-
-        if secure == None:
-            secure = _SECURE_DEFAULT
-        if secure:
-            self.jail = _Jail()
-            self.jail.prepare_with(working_directory)
-            self.working_directory = self.jail.home_dir
-        else:
-            self.jail = None
-            self.working_directory = working_directory
-
-    @property
-    def is_alive(self):
-        """Indicates whether a command is currently running in the sandbox"""
-        if self._is_alive:
-            sub_result = self.command_process.poll()
-            if sub_result is None:
-                return True
-            self._is_alive = False
-        return False
-
     def start(self, shell_command):
         """Start a command running in the sandbox"""
         if self.is_alive:
             raise SandboxError("Tried to run command with one in progress.")
-        if self.jail:
-            shell_command = self.jail.chroot_cmd + shell_command
-            working_directory = None
-        else:
-            working_directory = self.working_directory
+        shell_command = self.chroot_cmd + shell_command
         shell_command = shlex.split(shell_command.replace('\\','/'))
         try:
             self.command_process = subprocess.Popen(shell_command,
                                                     stdin=subprocess.PIPE,
                                                     stdout=subprocess.PIPE,
-                                                    stderr=subprocess.PIPE,
-                                                    cwd=working_directory)
+                                                    stderr=subprocess.PIPE,)
         except OSError:
             raise SandboxError('Failed to start {0}'.format(shell_command))
         self._is_alive = True
@@ -194,6 +163,15 @@ class Sandbox:
         stderr_monitor.daemon = True
         stderr_monitor.start()
 
+    def _signal(self, signal):
+        if not self.locked:
+            raise SandboxError("Attempt to send %s to unlocked jail" % (signal,))
+        result = subprocess.call("sudo -u {0} kill -{1} -1".format(
+            self.name, signal), shell=True)
+        if result != 0:
+            raise SandboxError("Error returned from jail %s sending signal %s"
+                    % (self.name, signal))
+
     def kill(self):
         """Stops the sandbox.
 
@@ -202,60 +180,16 @@ class Sandbox:
         suddenly terminated.
 
         """
-        if self.is_alive:
-            if self.jail:
-                self.jail.kill()
-            else:
-                try:
-                    self.command_process.kill()
-                except OSError:
-                    pass
-            self.command_process.wait()
-
-    def retrieve(self):
-        """Copy the working directory back out of the sandbox."""
-        if self.is_alive:
-            raise SandboxError("Tried to retrieve sandbox while still alive")
-        if self.jail:
-            self.jail.retrieve()
-
-    def release(self):
-        """Release the sandbox for further use
-
-        If running in a jail unlocks and releases the jail for reuse by others.
-        Must be called exactly once after Sandbox.kill has been called.
-
-        """
-        if self.is_alive:
-            raise SandboxError("Sandbox released while still alive")
-        if self.jail:
-            self.jail.release()
+        self._signal("KILL")
+        self._signal("CONT")
 
     def pause(self):
-        """Pause the process by sending a SIGSTOP to the child
-
-        This method is a no-op on Windows
-        """
-        if self.jail:
-            self.jail.pause()
-        else:
-            try:
-                self.command_process.send_signal(signal.SIGSTOP)
-            except (ValueError, AttributeError, OSError):
-                pass
+        """Pause the process by sending a SIGSTOP to the child"""
+        self._signal("STOP")
 
     def resume(self):
-        """Resume the process by sending a SIGCONT to the child
-
-        This method is a no-op on Windows
-        """
-        if self.jail:
-            self.jail.resume()
-        else:
-            try:
-                self.command_process.send_signal(signal.SIGCONT)
-            except (ValueError, AttributeError, OSError):
-                pass
+        """Resume the process by sending a SIGCONT to the child"""
+        self._signal("CONT")
 
     def write(self, str):
         """Write str to stdin of the process being run"""
@@ -314,6 +248,184 @@ class Sandbox:
             return self.stderr_queue.get(block=True, timeout=timeout)
         except Empty:
             return None
+
+
+class House:
+    """Provide an insecure sandbox to run arbitrary commands in.
+
+    The sandbox class is used to invoke arbitrary shell commands.
+    This class provides the same interface as the secure Sandbox but doesn't
+    provide any actual security or require any special system setup.
+
+    """
+
+    def __init__(self, working_directory):
+        """Initialize a new sandbox for the given working directory.
+
+        working_directory: the directory in which the shell command should
+                           be launched.
+        """
+        self._is_alive = False
+        self.command_process = None
+        self.stdout_queue = Queue()
+        self.stderr_queue = Queue()
+        self.working_directory = working_directory
+
+    @property
+    def is_alive(self):
+        """Indicates whether a command is currently running in the sandbox"""
+        if self._is_alive:
+            sub_result = self.command_process.poll()
+            if sub_result is None:
+                return True
+            self._is_alive = False
+        return False
+
+    def start(self, shell_command):
+        """Start a command running in the sandbox"""
+        if self.is_alive:
+            raise SandboxError("Tried to run command with one in progress.")
+        working_directory = self.working_directory
+        shell_command = shlex.split(shell_command.replace('\\','/'))
+        try:
+            self.command_process = subprocess.Popen(shell_command,
+                                                    stdin=subprocess.PIPE,
+                                                    stdout=subprocess.PIPE,
+                                                    stderr=subprocess.PIPE,
+                                                    cwd=working_directory)
+        except OSError:
+            raise SandboxError('Failed to start {0}'.format(shell_command))
+        self._is_alive = True
+        stdout_monitor = Thread(target=_monitor_file,
+                                args=(self.command_process.stdout, self.stdout_queue))
+        stdout_monitor.daemon = True
+        stdout_monitor.start()
+        stderr_monitor = Thread(target=_monitor_file,
+                                args=(self.command_process.stderr, self.stderr_queue))
+        stderr_monitor.daemon = True
+        stderr_monitor.start()
+
+    def kill(self):
+        """Stops the sandbox.
+
+        Stops down the sandbox, cleaning up any spawned processes, threads, and
+        other resources. The shell command running inside the sandbox may be
+        suddenly terminated.
+
+        """
+        if self.is_alive:
+            try:
+                self.command_process.kill()
+            except OSError:
+                pass
+            self.command_process.wait()
+
+    def retrieve(self):
+        """Copy the working directory back out of the sandbox."""
+        if self.is_alive:
+            raise SandboxError("Tried to retrieve sandbox while still alive")
+        pass
+
+    def release(self):
+        """Release the sandbox for further use
+
+        If running in a jail unlocks and releases the jail for reuse by others.
+        Must be called exactly once after Sandbox.kill has been called.
+
+        """
+        if self.is_alive:
+            raise SandboxError("Sandbox released while still alive")
+        pass
+
+    def pause(self):
+        """Pause the process by sending a SIGSTOP to the child
+
+        A limitation of the method is it will only pause the initial
+        child process created any further (grandchild) processes created
+        will not be paused.
+
+        This method is a no-op on Windows.
+        """
+        try:
+            self.command_process.send_signal(signal.SIGSTOP)
+        except (ValueError, AttributeError, OSError):
+            pass
+
+    def resume(self):
+        """Resume the process by sending a SIGCONT to the child
+
+        This method is a no-op on Windows
+        """
+        try:
+            self.command_process.send_signal(signal.SIGCONT)
+        except (ValueError, AttributeError, OSError):
+            pass
+
+    def write(self, str):
+        """Write str to stdin of the process being run"""
+        if not self.is_alive:
+            return False
+        try:
+            self.command_process.stdin.write(str)
+            self.command_process.stdin.flush()
+        except (OSError, IOError):
+            self.kill()
+            return False
+        return True
+
+    def write_line(self, line):
+        """Write line to stdin of the process being run
+
+        A newline is appended to line and written to stdin of the child process
+
+        """
+        if not self.is_alive:
+            return False
+        try:
+            self.command_process.stdin.write(line + "\n")
+            self.command_process.stdin.flush()
+        except (OSError, IOError):
+            self.kill()
+            return False
+        return True
+
+    def read_line(self, timeout=0):
+        """Read line from child process
+
+        Returns a line of the child process' stdout, if one isn't available
+        within timeout seconds it returns None. Also guaranteed to return None
+        at least once after each command that is run in the sandbox.
+
+        """
+        if not self.is_alive:
+            timeout=0
+        try:
+            return self.stdout_queue.get(block=True, timeout=timeout)
+        except Empty:
+            return None
+
+    def read_error(self, timeout=0):
+        """Read line from child process' stderr
+
+        Returns a line of the child process' stderr, if one isn't available
+        within timeout seconds it returns None. Also guaranteed to return None
+        at least once after each command that is run in the sandbox.
+
+        """
+        if not self.is_alive:
+            timeout=0
+        try:
+            return self.stderr_queue.get(block=True, timeout=timeout)
+        except Empty:
+            return None
+
+def get_sandbox(working_dir, secure=None):
+    if secure is None:
+        secure = server_info["secure_jail"]
+    if secure:
+        return Jail(working_dir)
+    else:
+        return House(working_dir)
 
 def main():
     parser = OptionParser(usage="usage: %prog [options] <command to run>")
